requirements.txt
streamlit
python-docx
streamlit_extras
